Renamed MosekInterface to MosekWrapper.
