+++ drake/solvers/CMakeLists.txt
@@ -49,7 +49,7 @@ else()
 endif()
 pods_find_pkg_config(mosek)
 if(mosek_FOUND)
-  list(APPEND optimization_files MosekSolver.cpp MosekInterface.cpp)
+  list(APPEND optimization_files MosekSolver.cpp MosekWrapper.cpp)
 else()
   list(APPEND optimization_files NoMosek.cpp)
 endif()
