+++ drake/solvers/MosekSolver.cpp
@@ -1,9 +1,9 @@
 // Copyright 2016, Alex Dunyak
-// A wrapper file for MosekInterface and mosekQP that handles constraint and
+// A wrapper file for MosekWrapper and mosekQP that handles constraint and
 // objective marshalling
 
 #include "drake/solvers/MosekSolver.h"
-#include "drake/solvers/MosekInterface.h"
+#include "drake/solvers/MosekWrapper.h"
 
 #include <Eigen/Core>
 
@@ -22,7 +22,7 @@ SolutionResult MosekSolver::Solve(OptimizationProblem& prog) const {
             != std::string::npos ||
         prog.GetSolverOptionsStr("Mosek").at("problemtype").find("quadratic")
             != std::string::npos) {
-      return MosekInterface::Solve(prog);
+      return MosekWrapper::Solve(prog);
     }
   }  // TODO(alexdunyak): add more mosek solution types.
   return kUnknownError;
